Let scrotwm adjust the font size on xterms as it squishes them in tiling mode to keep the terminal width above a certan size (set with 'term_width' in the config file.
We do this by sending the default font size change keystrokes to the xterm
in question. Because xterm does not accept 'synthetic' keystroke events by
default, and we don't want to force users to enable acceptSendEvents for
security reasons, hook XtAppNextEvent in the LD_PRELOAD hack, and clear
the send_event flag on the events in question. CAVEAT: Only works if xterm
is not setuid/setgid.