Hi, here's the patch for EWMH support for scrotwm. Again, I'm really sorry that it took this long. The patch is against the latest CVS version. Also, the code should adhere to style(9) this time.
Like I promised, the patch implements (at least) the following:
 - autoquirking using the _NET_WM_WINDOW_TYPE property (already present in
the previous patch)
 - the _NET_ACTIVE_WINDOW property should work properly now
 - fullscreen windows using the _NET_WM_STATE_FULLSCREEN property
 - floaters remember floating state and position when restarting scrotwm
 - implemented _NET_WM_STATE and _NET_WM_ALLOWED_ACTIONS.

There's one thing though. _NET_WM_ALLOWED_ACTIONS still allows only
floaters to be resized. This is because I couldn't think of a way to
implement the "resize the stack" in way that isn't surprinsing. For
example, when we have more than two columns (master and 1 other) in the
layout, I can't think of good way to react to a request for resizing a
window in the middle columns. I can implement it, but I need you to tell
me exactly how you want it to behave. :)

Do tell me if you find any bugs in the patch. All other feedback is
appreciated as well.

 - Tuukka

From: Tuukka Kataja <stuge@xor.fi>

Thanks and some cleanup by me.