Kill references before focusing on a new window.
Like the other focusing functions, make focus_prev call kill_refs before
trying to focus on a new window. This way, the current window references
are cleared like it should be when the very last window is destroyed.

Otherwise, the status-bar output will contain (while its alarm does not
expire) the class hint of the to be removed window.