Fix switchws issue when swapping with an empty workspace.
Fix screen switching issue when switching to a screen with an empty ws.

Improve root_to_region by adding check flags for pointer, focus, etc.

When moving mouse pointer to a region with an empty ws, focus on root.

Select EnterNotify events on root for moving mouse to an empty region.