Add new option: warp_focus Focus on the target window/workspace/region when clamped. For example, when attempting to switch to a workspace that is mapped on another region and workspace_clamp is enabled, focus on the region with the target workspace.  Enable by setting to 1.
Rename visible_noswap to workspace_clamp and update man page:
Prevents workspaces from being swapped when attempting to switch to a
workspace that is mapped to another region.  Use warp_focus if you want
to focus on the region containing the workspace and warp_pointer if you
want to also send the pointer.  Enable by setting to 1.

closes #80.