Add a workspace list to the bar
If '+L' appears in bar_format, display a list of workspaces associated
with the current screen. The set of workspaces shown can be:
a) all active workspaces (those containing windows or associated with
   the current region);
b) all active plus named workspaces (regardless of whether they are
   active); or
c) all workspaces.
The displayed subset is controlled via bar_ws_list_show.

A '*' character will precede the workspace of the current region.
Optionally, if bar_ws_list_urgent is set to true, a '!' character will
precede any workspace with an urgent window.

For example, suppose we have windows on workspaces 1 and 5, and we are
currently on workspace 4. Further, suppose workspace 1 is called "web"
and workspace 5 contains an urgent window (and bar_ws_list_urgent is
enabled). Then, the list would look like: [ 1:web *4 !5 ].