Added a config option to spawn programs on the current workspace.
If a programme was started and then new instances of that programme
was started while the original was still running the new instance would
end up in the original workspace.

By adding a new configuration option that unsets the environment
variable "_SWM_WS" the behaviour is changed to spawn new windows on
the workspace that is currently in focus.

The main idea for the fix came from forum post #3, by the member
"br", at https://opensource.conformal.com/fluxbb/viewtopic.php?pid=2569.
Other users in that thread have been using the code change without
problems and I have not noticed any side-effects.