CHANGELOG: Add hystorical releases
For each release, the date it taken from the corresponding
git tag; the one exception is 2.4.0, where the tag was created
months after the commit it points to.

Information about releases is taken from [1] and minimally
edited to work in the new context, along with being fixed when
necessary.

While every single release from 1.0.0 onwards is now accounted
for, several of them lack detailed information: tracking
existing release notes about them or writing them from scratch
is left as an exercise to the reader.

[1] https://sourceforge.net/projects/scrotwm/files/