Remove window border when there is only one window in workspace.
I would expect that no border is shown if there is only window in a
workspace. This is the case when disable_border is set, and the bar is
disabled.
Reading the manpage closely I concluded that my expectation is wrong:

disable_border
    Remove border when bar is disabled and there is only one
    window on the region."

I was wondering if it makes sense to change behaviour of
'disable_border' to:

    Remove border when bar is disabled OR there is only one window
    on the region.