Add support for _NET_WM_STATE_DEMANDS_ATTENTION.
Extends 'urgency hint' features with EWMH 'demands attention'.

Windows with the 'demands attention' hint are now recognized as 'urgent'
and can be quickly focused via the 'focus_urgent' action (default: M-u)

The 'Demands attention' state may be requested by client applications or
automatically set when a request to activate/focus a window is
unfulfilled (e.g. a program on another workspace requests focus on a new
tab/window that it opened.) It may be cleared via a request from client
applications or by focusing the window.

Closes #362