Add new `-d` command-line option to enable debug mode.
Debug mode and logging can now be enabled without the need of a rebuild!

When spectrwm is invoked with '-d', debug mode actions become available
and a debug log is printed to stderr.

For example, to capture a debug log when starting via .xinitrc:
  exec spectrwm -d > ~/spectrwm.log 2>&1

Debug mode also enables the following actions:
- 'debug_toggle' (Default: M-d)
  Toggles debug overlay.
- 'dumpwins' (Default: M-S-d)
  Prints current window/focus/stacking state to debug log.

Note that it is still possible to force-enable debug mode by building
with -DSWM_DEBUG.

An atom name cache is now used to avoid redundant GetAtomName requests
while printing the log.