do not display workspaces exceeding workspace_limit in the bar
This opens up the possibility to make a window completely invisible
by moving it out if the workspace_limit range, so there would be no
workspace showing up as active.

This is handy for scratchpads, that are shown on the current WS via
a keybinding and shall be completely hidden otherwise.