Drop Keywords from xsession configuration
I've added this a few years ago (9caab3f155d8) after Lintian
started complaining about it.

Looking back, that was probably not the right call: the
validation that is performed for .desktop files is based on the
assumption that they are describing applications, which makes
sense as that's the overwhelming majority of .desktop files out
there, but doesn't quite apply to us.

Looking at other xsession configurations shipped in Debian,
none of them includes the Keywords key. Let's drop it.