Add support for `+=` and `-=` assignment operators to the quirk option
Enables quirks to be assigned to windows using a combination of
different quirk entries.

When  a window is managed, quirk entries are applied in the order
specified in the configuration file. The assignment operator determines
how the quirks are applied. When assigning quirks with '=', quirks are
replaced on matching windows. To add or remove quirks, assign them with
'+=' or '-=' instead.

Example:
    quirk[.*] = FLOAT # Float all windows by default.
    quirk[XTerm] -= FLOAT # Don't float XTerm windows.
    quirk[.*:ws10] += WS[10] # Force workspace 10 and keep other quirks.
    quirk[Pcmanfm] += IGNORESPAWNWS # Add quirk to PCManFM windows.

Closes #107