Add new `spawn_flags` option
Change the spawn flags of program entry 'p'. If 'p' is omitted, the
default spawn flags are changed for any 'autorun' or 'program' entries
that come later in the configuration file.

One or more of the following flags may be specified in a
comma-separated list:
  nospawnws - Do not associate the program with the spawn workspace.
  xterm_fontadj - Enable automatic font size adjustments when resizing
                  xterm(1) windows that have the XTERM_FONTADJ quirk.
  optional - Disable program validation.
  none - No flags specified.

The default is 'none'.

In addition to the ‘=’ operator, this option also supports ‘+=’ and ‘-=’
to add/remove flags instead of replacing them.

Closes #565