Fix `_NET_WORKAREA` compatibility issue
It is common for clients to only look at the geometry of the first
desktop in the list and apply it regardless of the actual desktop its
window is on. This can result in menu positioning issues, etc. For
improved compatibility, list the workarea of the screen (root window)
for each desktop.