Add configuration reload
The configuration can now be reloaded without a restart. Note that this
also reinitializes the bars and can be used to restart bar_action
without the need of a full restart.

Add 'reload' action (default: unbound)
Reloads the configuration without restarting and reinitialize the bar(s)
along with bar_action (if used).

Add SIGUSR1 handling to initiate a reload.
Same as invoking the 'reload' action.