Add new quirk `NOTILE`
Disallow the window from being tiled. Note that the window will always
float regardless of whether _NET_WM_STATE_ABOVE is set.

This quirk is currently needed for floating Wine windows. When a Wine
window is floated, Wine requests the removal of _NET_WM_STATE_ABOVE,
which unfloats the window. Ignoring this request results in a
move/resize handling issue.